Add optional argument to LoadFromPostgreSql method
Added an optional argument, unknown, to the LoadFromPostgreSql method in the ReverseProxyBuilderExtensions class. This allows for more flexibility when invoking the method by enabling the handling of unknown properties. The changes also reflect in the Program.cs in the example directory to showcase the proper use of the newly added argument.