Update package versions and IDs
Updated the version number from "1.0.2" to "1.0.4" and changed the package ID from "Yarp.EfCore.Configuration.*" to "EfCore.Yarp.Configuration.*" in the Yarp.EfCore configuration files.

This change was necessary for the new release of our package and to align the package names with our current naming conventions for better consistency across the project.